backend/pyproject.toml
[tool.poetry]
name = "next-project-backend"
version = "0.1.0"
description = "Django backend for Next.js project"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.2.6"
djangorestframework = "^3.16.1"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
django-cors-headers = "^4.7.0"
dj-database-url = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.5.2"
black = "^23.0.0"
flake8 = "^6.0.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
manage = "manage:main"
